+++ PS7-x64/Modules/AdminToolbox.EndpointManagement/Public/Get-Applications.ps1
@@ -29,8 +29,6 @@ $computername = Get-Software
 
 .Link
 Uninstall-Application
-
-
 #>
 
     [OutputType('System.Software.Inventory')]