+++ PS7-x64/Modules/AdminToolbox.Exchange/Public/Get-MailboxScaleReport.ps1
@@ -29,6 +29,9 @@ Function Get-MailboxScaleReport {
     Must have the Exchange and importexcel modules loaded
     Exchange 2007: 50000 messages per folder, 500 folders total
     Exchange 2010/2013/2016/2019: 100000 messages per folder, 500 folders total
+
+    .LINK
+    Get-MailboxReport
     #>
 
     [CmdletBinding()]