+++ PS7-x64/Modules/AdminToolbox.Office365/Public/Set-Subscription.ps1
@@ -13,6 +13,7 @@ Function Set-Subscription {
 
     .LINK
     Get-AZContext
+    Set-ResourceGroup
     #>
 
     [CmdletBinding()]