+++ PS7-x64/Modules/AdminToolbox.Office365/ChangeLog.md
@@ -23,3 +23,4 @@
 * **1.3.0.1** Fixed errors in the get-office365 function
 * **2.0.0.0** Updated Get-Office365 and modified the logical structure of the module
 * **2.0.0.1** Applied fix to get-vmdetails
+* **2.1.0.0** Added Functions Get-TenantDiagrams and Invoke-EXOSharedSession